# api/openapi.yaml
openapi: 3.0.0
info:
  title: Portal OIDC
  version: "3.0"
  description: 認可認証基盤
  license:
    name: MIT
    url: "https://github.com/traPtitech/portal-oidc/blob/master/LICENSE"
  contact:
    name: traP
    url: "https://github.com/traPtitech/portal-oidc"
servers:
  - url: "http://localhost:8080"
    description: local

paths:
  /api/v1/admin/clients:
    get:
      operationId: getClients
      summary: クライアント一覧取得
      tags:
        - clients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          description: Unauthorized
    post:
      operationId: createClient
      summary: クライアント作成
      tags:
        - clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithSecret"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized

  /api/v1/admin/clients/{clientId}:
    parameters:
      - name: clientId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getClient
      summary: クライアント取得
      tags:
        - clients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
    put:
      operationId: updateClient
      summary: クライアント更新
      tags:
        - clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
    delete:
      operationId: deleteClient
      summary: クライアント削除
      tags:
        - clients
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found

  /api/v1/admin/clients/{clientId}/secret:
    parameters:
      - name: clientId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: regenerateClientSecret
      summary: クライアントシークレット再生成
      tags:
        - clients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientSecret"
        "401":
          description: Unauthorized
        "404":
          description: Not Found

  /.well-known/openid-configuration:
    get:
      operationId: getOpenIDConfiguration
      summary: OpenID Provider Configuration
      description: OpenID Connect Discovery 1.0 のメタデータを返却
      tags:
        - discovery
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenIDConfiguration"

  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      summary: JSON Web Key Set
      description: トークン検証用の公開鍵セットを返却
      tags:
        - discovery
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKS"

  /oauth2/introspect:
    post:
      operationId: introspectToken
      summary: トークンイントロスペクションエンドポイント (RFC 7662)
      description: |
        提示された access_token / refresh_token のメタデータを返却。
        active=false でアクティブでないことを示し、スコープ・有効期限・
        ユーザー識別子等を返す。クライアント認証必須。
      tags:
        - oauth2/oidc
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/IntrospectRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"

  /oauth2/userinfo:
    get:
      operationId: getUserInfo
      summary: UserInfo エンドポイント (GET)
      description: アクセストークンに紐づくユーザー情報を返却
      tags:
        - oauth2/oidc
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Unauthorized
    post:
      operationId: postUserInfo
      summary: UserInfo エンドポイント (POST)
      description: アクセストークンに紐づくユーザー情報を返却 (OIDC Core 1.0 準拠)
      tags:
        - oauth2/oidc
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Unauthorized

  /oauth2/authorize:
    get:
      operationId: getAuthorize
      summary: 認可エンドポイント (GET)
      description: OAuth2/OIDC認可リクエストを処理し、ユーザー認証後にリダイレクト (Query String Serialization)
      tags:
        - oauth2/oidc
      parameters:
        # OAuth 2.0 / OIDC 上は必須 (RFC 6749 Section 4.1.1, OIDC Core 1.0 Section 3.1.2.1)。
        # ただし欠落時のエラーは RFC 6749 Section 4.1.2.1 / OIDC Core 1.0 Section 3.1.2.6 に従い
        # 登録済み redirect_uri へ error として返さなければならず、素の HTTP 400 は仕様違反になる。
        # required: true にすると oapi-codegen 生成の binding がハンドラより先に 400 を返すため、
        # HTTP 層では optional として受け取り、必須判定は OAuth 層 (fosite) に委ねる。
        - name: response_type
          in: query
          required: false
          schema:
            type: string
            enum: [code]
          description: レスポンスタイプ (現在は code のみサポート。OAuth/OIDC 上は必須)
        - name: client_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: false
          schema:
            type: string
          description: スペース区切りのスコープ (openid, profile, email 等)
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: CSRF対策用のランダム文字列
        - name: nonce
          in: query
          required: false
          schema:
            type: string
          description: リプレイ攻撃対策用 (OIDC)
        - name: prompt
          in: query
          required: false
          schema:
            type: string
            enum: [none, login, consent, select_account]
          description: 認証・同意画面の表示制御 (OIDC Core 1.0 Section 3.1.2.1)
        - name: max_age
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: 最終認証からの最大許容経過秒数 (OIDC Core 1.0 Section 3.1.2.1)
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
          description: PKCE code_challenge (RFC 7636)
        - name: code_challenge_method
          in: query
          required: false
          schema:
            type: string
            enum: [S256, plain]
          description: PKCE code_challenge_method (RFC 7636)
      responses:
        "302":
          description: 認可成功時はredirect_uriへリダイレクト
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
    post:
      operationId: postAuthorize
      summary: 認可エンドポイント (POST)
      description: OAuth2/OIDC認可リクエストを処理し、ユーザー認証後にリダイレクト (Form Serialization)
      tags:
        - oauth2/oidc
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AuthorizeRequest"
      responses:
        "302":
          description: 認可成功時はredirect_uriへリダイレクト
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"

  /oauth2/token:
    post:
      operationId: token
      summary: トークンエンドポイント
      description: 認可コードをトークンに交換、またはトークンをリフレッシュ
      tags:
        - oauth2/oidc
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"

  /oauth2/revoke:
    post:
      operationId: revokeToken
      summary: トークン失効エンドポイント (RFC 7009)
      description: |
        指定された access_token または refresh_token を無効化する。
        RFC 7009 §2.2 により認可されたクライアントによる呼び出しは、
        トークンが既に存在しなかった場合でも 200 OK を返す。
      tags:
        - oauth2/oidc
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
      responses:
        "200":
          description: OK (トークンが無効化された / そもそも存在しなかった)
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"

components:
  schemas:
    AuthorizeRequest:
      type: object
      required:
        - response_type
        - client_id
        - redirect_uri
      properties:
        response_type:
          type: string
          enum: [code]
          description: レスポンスタイプ (現在は code のみサポート)
        client_id:
          type: string
          format: uuid
        redirect_uri:
          type: string
          format: uri
        scope:
          type: string
          description: スペース区切りのスコープ (openid, profile, email 等)
        state:
          type: string
          description: CSRF対策用のランダム文字列
        nonce:
          type: string
          description: リプレイ攻撃対策用 (OIDC)
        prompt:
          type: string
          enum: [none, login, consent, select_account]
          description: 認証・同意画面の表示制御 (OIDC Core 1.0 Section 3.1.2.1)
        max_age:
          type: integer
          minimum: 0
          description: 最終認証からの最大許容経過秒数 (OIDC Core 1.0 Section 3.1.2.1)
        code_challenge:
          type: string
          description: PKCE code_challenge (RFC 7636)
        code_challenge_method:
          type: string
          enum: [S256, plain]
          description: PKCE code_challenge_method (RFC 7636)

    ClientType:
      type: string
      enum:
        - public
        - confidential
      description: |
        - public: クライアントシークレットを安全に保持できないクライアント (SPA, モバイルアプリ等) <将来実装>
        - confidential: クライアントシークレットを安全に保持できるクライアント (サーバーサイドアプリ)

    Client:
      type: object
      required:
        - client_id
        - name
        - client_type
        - redirect_uris
        - created_at
        - updated_at
      properties:
        client_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        client_type:
          $ref: "#/components/schemas/ClientType"
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ClientCreate:
      type: object
      required:
        - name
        - client_type
        - redirect_uris
      properties:
        name:
          type: string
          maxLength: 255
        client_type:
          $ref: "#/components/schemas/ClientType"
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
          minItems: 1

    ClientUpdate:
      type: object
      required:
        - name
        - client_type
        - redirect_uris
      properties:
        name:
          type: string
          maxLength: 255
        client_type:
          $ref: "#/components/schemas/ClientType"
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
          minItems: 1

    ClientWithSecret:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          required:
            - client_secret
          properties:
            client_secret:
              type: string
              description: クライアントシークレット (作成時のみ返却、再取得不可)

    ClientSecret:
      type: object
      required:
        - client_secret
      properties:
        client_secret:
          type: string
          description: 新しいクライアントシークレット

    TokenRequest:
      type: object
      required:
        - grant_type
      properties:
        grant_type:
          type: string
          enum:
            - authorization_code
            - refresh_token
        code:
          type: string
          description: 認可コード (grant_type=authorization_code 時に必須)
        redirect_uri:
          type: string
          format: uri
          description: リダイレクトURI (grant_type=authorization_code 時に必須)
        client_id:
          type: string
          format: uuid
        client_secret:
          type: string
          description: クライアントシークレット (confidential client の場合)
        refresh_token:
          type: string
          description: リフレッシュトークン (grant_type=refresh_token 時に必須)
        code_verifier:
          type: string
          description: PKCE code_verifier

    IntrospectRequest:
      type: object
      description: |
        RFC 7662 §2.1 トークンイントロスペクションリクエスト。
      required:
        - token
      properties:
        token:
          type: string
          description: イントロスペクション対象のトークン
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
          description: トークン種別ヒント (RFC 7662 §2.1)
        client_id:
          type: string
          format: uuid
        client_secret:
          type: string

    IntrospectResponse:
      type: object
      description: |
        RFC 7662 §2.2 トークンイントロスペクションレスポンス。
        active=false の場合、他のフィールドは省略可能。
      required:
        - active
      properties:
        active:
          type: boolean
          description: トークンがアクティブか (RFC 7662 §2.2)
        scope:
          type: string
          description: スペース区切りのスコープ
        client_id:
          type: string
        username:
          type: string
        token_type:
          type: string
          description: 例 Bearer
        exp:
          type: integer
          format: int64
          description: トークン有効期限 (Unix epoch)
        iat:
          type: integer
          format: int64
          description: 発行時刻 (Unix epoch)
        nbf:
          type: integer
          format: int64
          description: 有効開始時刻 (Unix epoch)
        sub:
          type: string
          description: トークンサブジェクト (ユーザーID等)
        aud:
          type: array
          items:
            type: string
        iss:
          type: string
          description: 発行者 URL
        jti:
          type: string
          description: JWT ID

    RevokeRequest:
      type: object
      description: |
        RFC 7009 §2.1 トークン失効リクエスト。
      required:
        - token
      properties:
        token:
          type: string
          description: 失効させる access_token または refresh_token
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
          description: |
            トークン種別ヒント (RFC 7009 §2.1)。
            提供されると検索順序が最適化される。誤指定でも動作する。
        client_id:
          type: string
          format: uuid
        client_secret:
          type: string

    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: アクセストークンの有効期限 (秒)
        refresh_token:
          type: string
        id_token:
          type: string
          description: ID トークン (scope に openid が含まれる場合)
        scope:
          type: string
          description: 付与されたスコープ

    OAuthError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
            - server_error
        error_description:
          type: string
          description: エラーの詳細説明

    OpenIDConfiguration:
      type: object
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - jwks_uri
        - response_types_supported
        - subject_types_supported
        - id_token_signing_alg_values_supported
      properties:
        issuer:
          type: string
          format: uri
        authorization_endpoint:
          type: string
          format: uri
        token_endpoint:
          type: string
          format: uri
        userinfo_endpoint:
          type: string
          format: uri
        jwks_uri:
          type: string
          format: uri
        response_types_supported:
          type: array
          items:
            type: string
        subject_types_supported:
          type: array
          items:
            type: string
        id_token_signing_alg_values_supported:
          type: array
          items:
            type: string
        scopes_supported:
          type: array
          items:
            type: string
        token_endpoint_auth_methods_supported:
          type: array
          items:
            type: string
        claims_supported:
          type: array
          items:
            type: string
        code_challenge_methods_supported:
          type: array
          items:
            type: string
        request_parameter_supported:
          type: boolean
          description: |
            Indicates whether the OP supports use of the `request` parameter
            (OpenID Connect Core 1.0 §6 / Discovery 1.0 §3).
        request_uri_parameter_supported:
          type: boolean
          description: |
            Indicates whether the OP supports use of the `request_uri` parameter
            (OpenID Connect Core 1.0 §6 / Discovery 1.0 §3).
        require_request_uri_registration:
          type: boolean
          description: |
            Indicates whether the OP requires `request_uri` values to be
            pre-registered (OpenID Connect Discovery 1.0 §3).

    JWKS:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/JWK"

    JWK:
      type: object
      required:
        - kty
        - kid
        - use
      properties:
        kty:
          type: string
          description: Key Type (RSA, EC 等)
        kid:
          type: string
          description: Key ID
        use:
          type: string
          enum: [sig, enc]
        alg:
          type: string
          description: Algorithm (RS256 等)
        n:
          type: string
          description: RSA modulus (Base64url)
        e:
          type: string
          description: RSA exponent (Base64url)

    UserInfo:
      type: object
      required:
        - sub
      properties:
        sub:
          type: string
          description: Subject Identifier
        name:
          type: string
        preferred_username:
          type: string
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        picture:
          type: string
          format: uri
        updated_at:
          type: integer
          description: Unix timestamp

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer